Move the sketch.proto from the CMMS-api repo to anysketch. (#5)
* Move the sketch.proto from the CMMS-api repo to anysketch.

This file is copied from
https://github.com/world-federation-of-advertisers/cross-media-measurement-api/blob/c257ae4f0790d5080875edf80f536c4ae422ee19/src/main/proto/wfa/measurement/api/v2alpha/sketch.proto

The sketchConfig is acutally an anysketch concept. In CMMS, we use
protocol specific config + bytes to denote a sketch. Moving the
sketchproto into anysketch makes the ansketch repo a leaf repo.

The EncryptedSketch.proto and Sketch.proto inside CMMS-api would be
deleted.

* Drop the key from SketchConfig.

Updating proto field numbers is safe since the code is not used anywhere.